requirements.txt
aiocache==0.3.0
aiohttp==1.2.0
aiomcache==0.4.0
aioredis==0.3.0
appdirs==1.4.0
async-timeout==1.1.0
chardet==2.3.0
gunicorn==19.6.0
hiredis==0.2.0
multidict==2.1.4
packaging==16.8
pyparsing==2.1.10
PyYAML==3.12
six==1.10.0
uvloop==0.7.2
yarl==0.8.1
